server/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
certifi==2025.6.15
cffi==1.17.1
click==8.2.1
cryptography==45.0.4
dnspython==2.7.0
dotenv==0.9.9
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.13
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
motor==3.7.1
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pymongo==4.13.2
pytest==8.4.0
pytest-asyncio==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
